pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "ultra_ext"
version = "0.1.0"
description = "Extended utilities for YOLO and vision models"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Louis"}
]
keywords = ["yolo", "vision", "refcoco", "referring-expression", "object-detection"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "Pillow>=8.0.0",
    "requests>=2.25.0",
    "datasets>=2.0.0",
    "matplotlib>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.0",
    "flake8>=3.9",
]

[project.urls]
Homepage = "https://github.com/yourusername/ultra_ext"
Documentation = "https://github.com/yourusername/ultra_ext"
Repository = "https://github.com/yourusername/ultra_ext"

[tool.setuptools.packages.find]
include = ["ultra_ext*"]
exclude = ["tests*"]
